feature:create description box for org home page #1318
