+++ src/pages/tickets/org/orgHeader/index.tsx
@@ -1,9 +1,7 @@
 import React, { useState, useEffect, useRef } from 'react';
-import styled from 'styled-components';
 import { EuiCheckboxGroup, EuiPopover, EuiText } from '@elastic/eui';
 import MaterialIcon from '@material/react-material-icon';
 import { PostModal } from 'people/widgetViews/postBounty/PostModal';
-import { Organization } from 'store/main';
 import { GetValue, coding_languages } from 'people/utils/languageLabelStyle';
 import { colors } from 'config';
 import { OrgBountyHeaderProps } from '../../../../people/interfaces';
@@ -12,444 +10,46 @@ import addBounty from './Icons/addBounty.svg';
 import dropdown from './Icons/dropDownIcon.svg';
 import searchIcon from './Icons/searchIcon.svg';
 import file from './Icons/file.svg';
-import checkboxImage from './Icons/checkboxImage.svg';
 import githubIcon from './Icons/githubIcon.svg';
 import websiteIcon from './Icons/websiteIcon.svg';
+import {
+  BountyNumber,
+  Button,
+  CompanyDescription,
+  CompanyLabel,
+  CompanyNameAndLink,
+  DropDownButton,
+  EuiPopOverCheckbox,
+  EuiPopOverCheckboxRight,
+  FillContainer,
+  FilterLabel,
+  Filters,
+  FiltersLeft,
+  FiltersRight,
+  Header,
+  Icon,
+  ImageContainer,
+  Img,
+  InternalContainer,
+  Leftheader,
+  NewStatusContainer,
+  NumberOfBounties,
+  PrimaryText,
+  RightHeader,
+  SearchBar,
+  SearchWrapper,
+  SecondaryText,
+  SkillContainer,
+  SkillFilter,
+  SoryByContainer,
+  StatusContainer,
+  UrlButton,
+  UrlButtonContainer
+} from './OrgHeaderStyles';
+
+const color = colors['light'];
 
 const Coding_Languages = GetValue(coding_languages);
-interface styledProps {
-  color?: any;
-}
-const color = colors['light'];
-const Header = styled.div`
-  height: 130px;
-  width: 65%;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  background: #fff;
-`;
-
-const UrlButtonContainer = styled.div`
-  width: 180px;
-  display: flex;
-  gap: 8px;
-`;
-
-const FillContainer = styled.div`
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: #fff;
-  border-bottom: 1px solid var(--Input-BG-1, #f2f3f5);
-`;
-
-const Filters = styled.div`
-  display: flex;
-  width: 1366px;
-  padding: 10px 130px;
-  justify-content: center;
-  align-items: center;
-  gap: 198px;
-  align-self: stretch;
-  background: #fff;
-  margin-left: auto;
-  margin-right: auto;
-`;
-const FiltersRight = styled.span`
-  display: flex;
-  height: 40px;
-  padding-right: 122px;
-  align-items: flex-start;
-  gap: 52px;
-  flex: 1 0 0;
-  width: 1366px;
-`;
-
-const SkillContainer = styled.span`
-  padding: 10px 0px;
-  align-items: center;
-  display: flex;
-  position: relative;
-`;
-
-const Button = styled.button`
-  border-radius: 6px;
-  background: var(--Primary-Green, #49c998);
-  box-shadow: 0px 2px 10px 0px rgba(73, 201, 152, 0.5);
-  border: none;
-  display: flex;
-  width: 144px;
-  height: 40px;
-  padding: 8px 16px;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 6px;
-  color: var(--White, #fff);
-  text-align: center;
-  font-family: Barlow;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 0px; /* 0% */
-  letter-spacing: 0.14px;
-`;
-
-const UrlButton = styled.button`
-  border-radius: 4px;
-  margin-right: auto;
-  border: 1px solid #dde1e5;
-  background: #ffffff;
-  display: flex;
-  width: 85px;
-  height: 28px;
-  padding: 8px 16px;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-  color: var(--Black, #5f6368);
-  text-align: center;
-  font-family: 'Barlow';
-  font-size: 13px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 0px; /* 0% */
-  letter-spacing: 0.14px;
-`;
-
-const FilterLabel = styled.label`
-  color: var(--Main-bottom-icons, #5f6368);
-  font-family: Barlow;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 17px;
-  letter-spacing: 0.15px;
-`;
-const Leftheader = styled.div`
-  display: flex;
-  gap: 20px;
-  align-items: center;
-  justify-content: center;
-`;
-const CompanyNameAndLink = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
-
-const CompanyLabel = styled.p`
-  color: var(--Text-2, #3c3f41);
-  font-family: Barlow;
-  font-size: 28px;
-  font-style: normal;
-  font-weight: 700;
-`;
-const ImageContainer = styled.img`
-  width: 72px;
-  height: 72px;
-  border-radius: 66px;
-`;
-
-const CompanyDescription = styled.div`
-  max-width: 403px;
-  text-align: right;
-  color: var(--Main-bottom-icons, #5f6368);
-  font-family: Barlow;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-`;
-const RightHeader = styled.div`
-  max-width: 593px;
-  display: flex;
-  gap: 46px;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const SearchWrapper = styled.div`
-  height: 40px;
-  padding: 0px 16px;
-  align-items: center;
-  gap: 10px;
-  flex: 1 0 0;
-  display: flex;
-  position: relative;
-`;
-
-const Icon = styled.img`
-  position: absolute;
-  right: 30px;
-`;
-
-const SearchBar = styled.input`
-  display: flex;
-  height: 40px;
-  padding: 0px 16px;
-  padding-left: 30px;
-  align-items: center;
-  gap: 10px;
-  flex: 1 0 0;
-  border-radius: 6px;
-  background: var(--Input-BG-1, #f2f3f5);
-  outline: none;
-  border: none;
-`;
-
-const SoryByContainer = styled.span`
-  padding: 10px 0px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const NumberOfBounties = styled.div`
-  height: 23px;
-  padding: 1.5px 983.492px 1.5px 10px;
-  align-items: center;
-  flex-shrink: 0;
-  margin: 23px 133px;
-  width: 1366px;
-  margin-left: auto;
-  margin-right: auto;
-`;
-const BountyNumber = styled.span`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 11px;
-`;
-const PrimaryText = styled.p`
-  color: var(--Primary-Text-1, var(--Press-Icon-Color, #292c33));
-  font-family: Barlow;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-`;
-const SecondaryText = styled.p`
-  color: var(--Main-bottom-icons, #5f6368);
-  font-family: Barlow;
-  font-size: 15px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-`;
-const Img = styled.img`
-  padding-bottom: 10px;
-`;
-
-const SkillFilter = styled.div`
-  width: 480px;
-  height: 280px;
-  background-color: white;
-  position: absolute;
-  top: 50px;
-  z-index: 999;
-  border-radius: 0px 0px 6px 6px;
-  border-right: 1px solid rgba(0, 0, 0, 0.1);
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  border-left: 1px solid rgba(0, 0, 0, 0.1);
-  background: #fff;
-  box-shadow: 0px 2px 8px 0px rgba(0, 0, 0, 0.12);
-
-  /* border-top: 3px solid var(--Primary-blue, #618AFF);
-  border-top-height: 20px; */
-
-  ::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    right: 380px;
-    top: 0;
-    height: 3px;
-    background: var(--Primary-blue, #618aff);
-  }
-`;
-const InternalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 30px;
-  padding: 24px 20px 28px 20px;
-`;
-
-const EuiPopOverCheckboxRight = styled.div<styledProps>`
-  height: auto;
-  user-select: none;
-  &.CheckboxOuter > div {
-    height: 100%;
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    column-gap: 56px;
-    justify-content: center;
-    .euiCheckboxGroup__item {
-      .euiCheckbox__square {
-        top: 5px;
-        border: 1px solid ${(p: any) => p?.color && p?.color?.grayish.G500};
-        border-radius: 2px;
-      }
-      .euiCheckbox__input + .euiCheckbox__square {
-        background: ${(p: any) => p?.color && p?.color?.pureWhite} no-repeat center;
-      }
-      .euiCheckbox__input:checked + .euiCheckbox__square {
-        border: 1px solid ${(p: any) => p?.color && p?.color?.blue1};
-        background: ${(p: any) => p?.color && p?.color?.blue1} no-repeat center;
-        background-image: url('static/checkboxImage.svg');
-      }
-      .euiCheckbox__label {
-        font-family: 'Barlow';
-        font-style: normal;
-        font-weight: 500;
-        font-size: 13px;
-        line-height: 18px;
-        color: ${(p: any) => p?.color && p?.color?.grayish.G50};
-        &:hover {
-          color: ${(p: any) => p?.color && p?.color?.grayish.G05};
-        }
-      }
-      input.euiCheckbox__input:checked ~ label {
-        color: ${(p: any) => p?.color && p?.color?.blue1};
-      }
-    }
-  }
-`;
-
-const DropDownButton = styled.button`
-  border: none;
-  background-color: transparent;
-  padding-top: 5px;
-`;
-
-const FiltersLeft = styled.span`
-  display: flex;
-  height: 40px;
-  align-items: flex-start;
-`;
-
-const EuiPopOverCheckbox = styled.div<styledProps>`
-  width: 147px;
-  height: auto;
-  padding: 15px 18px;
-  border-right: 1px solid ${(p: any) => p.color && p.color.grayish.G700};
-  user-select: none;
-  .leftBoxHeading {
-    font-family: 'Barlow';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 12px;
-    line-height: 32px;
-    text-transform: uppercase;
-    color: ${(p: any) => p.color && p.color.grayish.G100};
-    margin-bottom: 10px;
-  }
-
-  &.CheckboxOuter > div {
-    display: flex;
-    flex-direction: column;
-    flex-wrap: wrap;
-
-    .euiCheckboxGroup__item {
-      .euiCheckbox__square {
-        top: 5px;
-        border: 1px solid ${(p: any) => p?.color && p?.color?.grayish.G500};
-        border-radius: 2px;
-      }
-      .euiCheckbox__input + .euiCheckbox__square {
-        background: ${(p: any) => p?.color && p?.color?.pureWhite} no-repeat center;
-      }
-      .euiCheckbox__input:checked + .euiCheckbox__square {
-        border: 1px solid ${(p: any) => p?.color && p?.color?.blue1};
-        background: ${(p: any) => p?.color && p?.color?.blue1} no-repeat center;
-        background-size: contain;
-        background-image: url(${checkboxImage});
-      }
-      .euiCheckbox__label {
-        font-family: 'Barlow';
-        font-style: normal;
-        font-weight: 500;
-        font-size: 13px;
-        line-height: 16px;
-        color: ${(p: any) => p?.color && p?.color?.grayish.G50};
-        &:hover {
-          color: ${(p: any) => p?.color && p?.color?.grayish.G05};
-        }
-      }
-      input.euiCheckbox__input:checked ~ label {
-        color: ${(p: any) => p?.color && p?.color?.grayish.G05};
-      }
-    }
-  }
-`;
-
-const NewStatusContainer = styled.div``;
-
-const StatusContainer = styled.div<styledProps>`
-  width: 70px;
-  height: 48px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  margin-left: 19px;
-  margin-top: 4px;
-  cursor: pointer;
-  user-select: none;
-  .filterStatusIconContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 48px;
-    width: 38px;
-    .materialStatusIcon {
-      color: ${(p: any) => p.color && p.color.grayish.G200};
-      cursor: pointer;
-      font-size: 18px;
-      margin-top: 5px;
-    }
-  }
-  .statusText {
-    font-family: 'Barlow';
-    font-style: normal;
-    font-weight: 500;
-    font-size: 15px;
-    line-height: 17px;
-    letter-spacing: 0.15px;
-    display: flex;
-    align-items: center;
-    color: ${(p: any) => p.color && p.color.grayish.G200};
-  }
-  &:hover {
-    .filterStatusIconContainer {
-      .materialStatusIcon {
-        color: ${(p: any) => p.color && p.color.grayish.G50} !important;
-        cursor: pointer;
-        font-size: 18px;
-        margin-top: 5px;
-      }
-    }
-    .statusText {
-      color: ${(p: any) => p.color && p.color.grayish.G50};
-    }
-  }
-  &:active {
-    .filterStatusIconContainer {
-      .materialStatusIcon {
-        color: ${(p: any) => p.color && p.color.grayish.G10} !important;
-        cursor: pointer;
-        font-size: 18px;
-        margin-top: 5px;
-      }
-    }
-    .statusText {
-      color: ${(p: any) => p.color && p.color.grayish.G10};
-    }
-  }
-`;
 
 const Status = ['Open', 'Assigned', 'Completed', 'Paid'];
 export const OrgHeader = ({
@@ -480,11 +80,11 @@ export const OrgHeader = ({
   const handlePostBountyClose = () => {
     setIsPostBountyModalOpen(false);
   };
-  const handleWebsiteButton = (websiteUrl?: string) => {
+  const handleWebsiteButton = (websiteUrl: string) => {
     window.open(websiteUrl, '_blank');
   };
 
-  const handleGithubButton = (githubUrl?: string) => {
+  const handleGithubButton = (githubUrl: string) => {
     window.open(githubUrl, '_blank');
   };
 
@@ -499,9 +99,7 @@ export const OrgHeader = ({
     }
   }, [org_uuid, checkboxIdToSelectedMap, main, languageString]);
 
-  if (!organizationData) return null;
-
-  const { name, img, description, website, github } = organizationData as Organization;
+  const { name, img, description, website, github } = organizationData;
 
   const handleClick = () => {
     setFilterClick(!filterClick);
@@ -533,7 +131,7 @@ export const OrgHeader = ({
               <CompanyLabel>{name}</CompanyLabel>
               <UrlButtonContainer data-testid="url-button-container">
                 {website !== '' ? (
-                  <UrlButton onClick={() => handleWebsiteButton(website)}>
+                  <UrlButton onClick={() => handleWebsiteButton(website ?? '')}>
                     <img src={websiteIcon} alt="" />
                     Website
                   </UrlButton>
@@ -541,7 +139,7 @@ export const OrgHeader = ({
                   ''
                 )}
                 {github !== '' ? (
-                  <UrlButton onClick={() => handleGithubButton(github)}>
+                  <UrlButton onClick={() => handleGithubButton(github ?? '')}>
                     <img src={githubIcon} alt="" />
                     Github
                   </UrlButton>
@@ -628,7 +226,6 @@ export const OrgHeader = ({
             <SkillContainer>
               <FilterLabel>Skill</FilterLabel>
               <DropDownButton onClick={handleClick} data-testid="skillDropdown">
-                {' '}
                 <Img src={dropdown} alt="" />
               </DropDownButton>
               {filterClick ? (
@@ -656,7 +253,6 @@ export const OrgHeader = ({
             <SoryByContainer>
               <FilterLabel>Sort by:Newest First</FilterLabel>
               <DropDownButton>
-                {' '}
                 <Img src={dropdown} alt="" />
               </DropDownButton>
             </SoryByContainer>
